Implement 2 more UMI correction methods: adjacency and directional
